button: move component logic to new file
